add features showProduct and createcateory
