+++ frontend_nustly/src/features/Catalog/CreateCatalog/index.ts
@@ -1,0 +1,1 @@
+export { default as CreateCatalog } from './ui/CreateCatalog.vue';
