+++ frontend_nustly/src/features/Product/ui/ShowProduct/ShowProduct.vue
@@ -1,0 +1,11 @@
+<template>
+  <div>
+
+  </div>
+</template>
+
+<script setup lang="ts">
+
+</script>
+
+<style src="./ShowProduct.style.scss" lang="scss" scoped></style>
