+++ frontend_nustly/src/features/Product/ui/index.ts
@@ -1,0 +1,1 @@
+export { default as ShowProduct } from './ShowProduct/ShowProduct.vue'
